LayoutTests/imported/w3c/web-platform-tests/streams/readable-byte-streams/brand-checks.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

Harness Error (FAIL), message = ReferenceError: Can't find variable: SharedWorker
